fix(perfil): Corrige ordenação por data de submissões do perfil
