+++ views/index.pug
@@ -28,7 +28,7 @@ block content
                 li(style="font-weight: 600;") Melhorias nas submissões
                 li(style="font-weight: 600;") Abrir uma nova janela com as funções disponíveis
               br
-              p.has-text-justify Se você gosta de programação e deseja ajudar de alguma forma, entre em contato com Gabriel pelo email: 
+              p.has-text-justify Se você gosta de programação web e deseja ajudar de alguma forma, entre em contato com Gabriel pelo email: 
                 b tibuurcio@gmail.com
               p.has-text-justify Estamos trabalhando para adicionar novas questões, questões de desafio e questões de provas anteriores, entretanto esse é um trabalho que leva um pouco de tempo. Ainda há muito o que fazer, por isso precisamos da sua ajuda! 👍 
               br
